refactor: extract checkpoint logic to file
